Fix: refractor folder structure and rename the pathname in controllers
