+++ src/main/java/ca/openricecan/controller/DishController.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "dish")
+@RequestMapping(path = "api/dish")
 public class DishController {
     private final DishService dishService;
