+++ src/main/java/ca/openricecan/controller/PaymentMethodController.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "payment-method")
+@RequestMapping(path = "api/payment-method")
 public class PaymentMethodController {
     private final PaymentMethodService paymentMethodService;
