+++ src/main/java/ca/openricecan/controller/RestaurantOwnerController.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "restaurant-owner")
+@RequestMapping(path = "api/restaurant/owner")
 public class RestaurantOwnerController {
     private final RestaurantOwnerService restaurantOwnerService;
