+++ src/main/java/ca/openricecan/controller/ReviewController.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "review")
+@RequestMapping(path = "api/review")
 public class ReviewController {
     private final ReviewService reviewService;
