+++ src/main/java/ca/openricecan/controller/SubscribeController.java
@@ -1,14 +1,14 @@
 package ca.openricecan.controller;
 
-import ca.openricecan.data.entity.restaurant.SubscribeEntity;
-import ca.openricecan.service.restaurant.SubscribeService;
+import ca.openricecan.data.entity.subscribe.SubscribeEntity;
+import ca.openricecan.service.subscribe.SubscribeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(path = "subscribe")
+@RequestMapping(path = "api/subscribe")
 public class SubscribeController {
     private final SubscribeService subscribeService;
