+++ src/main/java/ca/openricecan/controller/UserController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(path = "user")
+@RequestMapping(path = "api/user")
 public class UserController {
     private final UserService userService;
