+++ src/main/java/ca/openricecan/service/subscribe/SubscribeService.java
@@ -1,7 +1,7 @@
-package ca.openricecan.service.restaurant;
+package ca.openricecan.service.subscribe;
 
-import ca.openricecan.data.entity.restaurant.SubscribeEntity;
-import ca.openricecan.repository.restaurant.SubscribeRepository;
+import ca.openricecan.data.entity.subscribe.SubscribeEntity;
+import ca.openricecan.repository.subscribe.SubscribeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
